Move all Grader classes to same location
