+++ src/main/java/edu/byu/cs/autograder/Grader.java
@@ -5,8 +5,8 @@ import edu.byu.cs.autograder.database.DatabaseHelper;
 import edu.byu.cs.autograder.git.CommitVerificationConfig;
 import edu.byu.cs.autograder.git.CommitVerificationResult;
 import edu.byu.cs.autograder.git.GitHelper;
-import edu.byu.cs.autograder.git.GitHubAssignmentGrader;
-import edu.byu.cs.autograder.quality.QualityGrader;
+import edu.byu.cs.autograder.test.GitHubAssignmentGrader;
+import edu.byu.cs.autograder.test.QualityGrader;
 import edu.byu.cs.autograder.score.Scorer;
 import edu.byu.cs.autograder.test.PassoffTestGrader;
 import edu.byu.cs.autograder.test.PreviousPhasePassoffTestGrader;
